お気に入り登録、解除機能のアップデート

+++ app/Http/Controllers/User/FavoriteController.php
@@ -26,11 +26,35 @@ class FavoriteController extends Controller
     //     //return back();
     // }
 
-    public function destroy(Request $request)
-    {
-        $favorite = Favorite::where('topic_id', $request->topic_id)->first();
+    // public function destroy(Request $request)
+    // {
+    //     $favorite = Favorite::where('user_id', Auth::id())
+    //                         ->where('topic_id', $request->topic_id)
+    //                         ->first();
         
-        $favorite->delete();
+    //     $favorite->delete();
+        
+    //     //return redirect('user/topic/browse?id='. $request->topic_id);
+    //     return back();
+    // }
+    
+    public function toggle(Request $request)
+    {
+        $favorite = Favorite::where('user_id', Auth::id())
+                            ->where('topic_id', $request->topic_id)
+                            ->first();
+        if (empty($favorite)){
+            //お気に入り未登録
+            //お気に入りを登録する
+            $favorite = new Favorite;
+            $favorite->topic_id = $request->topic_id;
+            $favorite->user_id = Auth::id();
+            $favorite->save();
+        }else{
+            //お気に入り登録済み
+            //お気に入りを削除する
+            $favorite->delete();
+        }
         
         //return redirect('user/topic/browse?id='. $request->topic_id);
         return back();
